requirements.txt
discord.py>=2.3.2
PyNaCl>=1.5.0
